// src/App/types/TimeCardTypes/BodyData/HollowTime.ts
export type HollowTime = {
  goingOutTimeCalcName: string
  goingOutTimeCalcValue: string
  // 外出時刻
  goingOutTimeHoursName: string
  goingOutTimeHoursValue: string
  goingOutTimeMinutesName: string
  goingOutTimeMinutesValue: string
  returnTimeCalcName: string
  returnTimeCalcValue: string
  returnTimeErrorName: string
  returnTimeErrorValue: string
  // 戻り時刻
  returnTimeHoursName: string
  returnTimeHoursValue: string
  returnTimeMinutesName: string
  returnTimeMinutesValue: string
}