// src/App/types/TimeCardTypes/BodyData/WorkTimes.ts
export type WorkTimes = {
  // 休憩時刻
  breakTimeCalcName: string
  breakTimeCalcValue: string
  breakTimeHoursName: string
  breakTimeHoursValue: string
  breakTimeMinutesName: string
  breakTimeMinutesValue: string
  // 退社時刻
  leaveTimeCalcName: string
  leaveTimeCalcValue: string
  leaveTimeHoursName: string
  leaveTimeHoursValue: string
  leaveTimeMinutesName: string
  leaveTimeMinutesValue: string
  // 出社時刻
  timeOfArrivalCalcName: string
  timeOfArrivalCalcValue: string
  timeOfArrivalHoursName: string
  timeOfArrivalHoursValue: string
  timeOfArrivalMinutesName: string
  timeOfArrivalMinutesValue: string
}